Moving Tests to own folder
