+++ extensions/docker-publisher/src/backend/__tests__/DockerHubAPI.test.ts
@@ -3,10 +3,10 @@
 
 import { fetch } from 'node-fetch';
 
-import { IRepository } from '../types/interfaces';
-import { RepositoryAPIProps } from '../types';
+import { IRepository } from '../../types/interfaces';
+import { RepositoryAPIProps } from '../../types';
 
-import { DockerHubAPI } from './DockerHubAPI';
+import { DockerHubAPI } from '../DockerHubAPI';
 
 jest.mock('node-fetch', () => ({
   fetch: jest.fn(),
